Created isDir function
Moved all functions that use ctx to context.go

Created a test helper for `needsGit` and `needsExternalNetwork`

Added a test case for `absoluteProjectPath` to make sure a file will error out

Renamed `getSourceManager` to `sourceManager`
